VideoClipper/requirements.txt
yt-dlp
moviepy
ffmpeg-python
openai-whisper
gTTS